tech(TTB-060): Renamed common directory to components
